Secure application by applying permissions
Most of the application is limited to eXist users in the dba group.
The activity log and the catalog are completely unreadable outside
that group in order to protect their data.

The rescheduler module is executable by any user. Doing so will
change the user's effective group to dba, allowing limited access
to the rest of the codebase for the purposes of rescheduling jobs.